% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS.R
\name{CLASS}
\alias{CLASS}
\title{CLASS}
\usage{
CLASS(
  X = NULL,
  y = NULL,
  csv = NULL,
  header = FALSE,
  nSample = -1,
  nTimes = -1,
  k = -1
)
}
\arguments{
\item{X}{Numeric matrix of predictors (without intercept*).}

\item{y}{Numeric vector of response variable.}

\item{csv}{String; path to the data (csv file).}

\item{header}{Logical; whether the csv files contains a header}

\item{nSample}{Integer; sample size of the uniform subsample done in CLASS.}

\item{nTimes}{Integer; number of times/iterations of LASSO in CLASS.}

\item{k}{Integer; number of rows in the subselection using IBOSS.}
}
\value{
A list with:
\itemize{
}
}
\description{
Runs the CLASS algorithm on given dataset.
}
