% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSRHT.R
\name{FeatureSRHT}
\alias{FeatureSRHT}
\title{FeatureSRHT: Subsampled Randomized Hadamard Transform for OLS}
\usage{
FeatureSRHT(X, y, r, method = "uniform", bins = 10, alpha = 0)
}
\arguments{
\item{X}{Numeric matrix of features}

\item{y}{Numeric vector of targets}

\item{r}{Integer, number of features to select}

\item{method}{String: 'uniform', 'top-r', 'leverage', 'supervised', or 'all'}

\item{bins}{Integer, number of quantile bins for supervised metric (default 10)}

\item{alpha}{Double, weight for intra-class variance penalty (default 0.0)}
}
\description{
FeatureSRHT: Subsampled Randomized Hadamard Transform for OLS
}
